1237-1238(1227)

+++ protected/models/Simulations.php
@@ -130,6 +130,9 @@ class Simulations extends CActiveRecord
             if (preg_match('/^M/', $event_trigger->event_sample->code)) {
                 continue;
             }
+            if ($event_trigger->trigger_time == '00:00:00') {
+                continue;
+            }
             if (GameTime::timeToSeconds($event_trigger->trigger_time) < ($newHours*60 + $newMinutes)*60) {
                 $event_trigger->delete();
             }
